+++ selfdrive/controls/lib/longitudinal_mpc_lib/long_mpc.py
@@ -8,6 +8,8 @@ from selfdrive.swaglog import cloudlog
 from selfdrive.modeld.constants import index_function
 from selfdrive.controls.lib.radar_helpers import _LEAD_ACCEL_TAU
 from selfdrive.car.modules.CFG_module import load_float_param
+from selfdrive.car.tesla.values import TESLA_MIN_ACCEL
+
 
 if __name__ == '__main__':  # generating code
   from pyextra.acados_template import AcadosModel, AcadosOcp, AcadosOcpSolver
@@ -49,17 +51,19 @@ T_IDXS_LST = [index_function(idx, max_val=MAX_T, max_idx=N+1) for idx in range(N
 
 T_IDXS = np.array(T_IDXS_LST)
 T_DIFFS = np.diff(T_IDXS, prepend=[0.])
-MIN_ACCEL = -4.5
+MIN_ACCEL = TESLA_MIN_ACCEL #was -3.5, fewer places to tune
 T_FOLLOW = load_float_param("TinklaFollowDistance",1.45)
 COMFORT_BRAKE = 2.5
-STOP_DISTANCE = 6.0 #base it if we have radar or just visual
+STOP_DISTANCE = 3.0 #base it if we have radar or just visual
 DIST_FACTOR = 2
+DIST_FACTOR_STOPPED = 4
+V_EGO_D = 1.
 
 def get_stopped_equivalence_factor(v_lead):
-  return (v_lead**2) / (DIST_FACTOR * COMFORT_BRAKE)
+  return (v_lead**2) / (DIST_FACTOR_STOPPED * COMFORT_BRAKE)
 
 def get_safe_obstacle_distance(v_ego,t_follow=T_FOLLOW):
-  return (v_ego**2) / (DIST_FACTOR * COMFORT_BRAKE) + t_follow * v_ego + STOP_DISTANCE
+  return (v_ego**2) / (DIST_FACTOR_STOPPED * COMFORT_BRAKE) + t_follow * v_ego + STOP_DISTANCE
   
 
 def desired_follow_distance(v_ego, v_lead,t_follow=T_FOLLOW):
@@ -137,7 +141,7 @@ def gen_long_mpc_solver():
   # from an obstacle at every timestep. This obstacle can be a lead car
   # or other object. In e2e mode we can use x_position targets as a cost
   # instead.
-  costs = [((x_obstacle - x_ego) - (desired_dist_comfort)) / (v_ego + 10.),
+  costs = [((x_obstacle - x_ego) - (desired_dist_comfort)) / (v_ego + V_EGO_D),
            x_ego,
            v_ego,
            a_ego,
@@ -152,7 +156,7 @@ def gen_long_mpc_solver():
   constraints = vertcat(v_ego,
                         (a_ego - a_min),
                         (a_max - a_ego),
-                        ((x_obstacle - x_ego) - (3/4) * (desired_dist_comfort)) / (v_ego + 10.))
+                        ((x_obstacle - x_ego) - (3/4) * (desired_dist_comfort)) / (v_ego + V_EGO_D))
   ocp.model.con_h_expr = constraints
   ocp.model.con_h_expr_e = vertcat(np.zeros(CONSTR_DIM))
 
@@ -310,7 +314,7 @@ class LongitudinalMpc:
     v_ego = self.x0[1]
     a_ego = self.x0[2]
     if carstate.followDistanceS != 255:
-      self.t_follow = 0.7 + float(carstate.followDistanceS) * 0.1
+      self.t_follow = 0.7 + float(carstate.followDistanceS) * 0.2
     self.status = radarstate.leadOne.status or radarstate.leadTwo.status
 
     lead_xv_0 = self.process_lead(radarstate.leadOne)