+++ RELEASES.md
@@ -1,11 +1,17 @@
-Unity Version 0.8.13-56 (2022-10-17)
+Unity Version 0.8.13-56 (2022-10-27)
 ========================
  * Sound files volume fix
- * Show steering icon on IC for AP cars when OP is available
+ * Show TACC icon on IC for AP cars when OP is available
  * Display shutdown
  * Engage at 0 MPH (AP cars)
  * Fleet based speed (AP cars)
- * Slow down in turns
+ * Slow down in turns based on SunnyPilot 
+ * prevent iBooster from pressing both pedals
+ * improved acceleration handling for AP1
+ * Add message ID for the CAN Error 
+ * Reduce min accel to -4.5 
+ * Improve follow distance
+ * Better and smoother pedal with 4 proiles
 
 Unity Version 0.8.13-55 (2022-09-21)
 ========================